# perf(html/parser): improve