--- a/crates/swc_html_parser/benches/lexer.rs
+++ b/crates/swc_html_parser/benches/lexer.rs
@@ -4,7 +4,7 @@ use criterion::{black_box, criterion_group, criterion_main, Bencher, Criterion};
 use swc_common::{input::StringInput, FileName};
 use swc_html_parser::lexer::Lexer;
 
-fn bench_lexer(b: &mut Bencher, src: &'static str) {
+fn bench_document(b: &mut Bencher, src: &'static str) {
     let _ = ::testing::run_test(false, |cm, _| {
         let fm = cm.new_source_file(FileName::Anon, src.into());
 
@@ -20,18 +20,30 @@ fn bench_lexer(b: &mut Bencher, src: &'static str) {
     });
 }
 
+fn run(c: &mut Criterion, id: &str, src: &'static str) {
+    c.bench_function(&format!("html/lexer/{}", id), |b| {
+        bench_document(b, src);
+    });
+}
+
 fn bench_files(c: &mut Criterion) {
-    c.bench_function("html/lexer/css_2021_spec", |b| {
-        bench_lexer(b, include_str!("./files/css_2021_spec.html"))
-    });
+    run(
+        c,
+        "css_2021_spec",
+        include_str!("./files/css_2021_spec.html"),
+    );
 
-    c.bench_function("html/lexer/github_com_17_05_2022", |b| {
-        bench_lexer(b, include_str!("./files/github_com_17_05_2022.html"))
-    });
+    run(
+        c,
+        "github_com_17_05_2022",
+        include_str!("./files/github_com_17_05_2022.html"),
+    );
 
-    c.bench_function("html/lexer/stackoverflow_com_17_05_2022", |b| {
-        bench_lexer(b, include_str!("./files/stackoverflow_com_17_05_2022.html"))
-    });
+    run(
+        c,
+        "stackoverflow_com_17_05_2022",
+        include_str!("./files/stackoverflow_com_17_05_2022.html"),
+    );
 }
 
 criterion_group!(benches, bench_files);