--- a/crates/swc_html_parser/src/lexer/mod.rs
+++ b/crates/swc_html_parser/src/lexer/mod.rs
@@ -225,9 +225,9 @@ where
     fn validate_input_stream_character(&mut self, c: char) {
         let code = c as u32;
 
-        if (0xd800..=0xdfff).contains(&code) {
+        if is_surrogate(code) {
             self.emit_error(ErrorKind::SurrogateInInputStream);
-        } else if code != 0x00 && is_control(code) {
+        } else if is_allowed_control_character(code) {
             self.emit_error(ErrorKind::ControlCharacterInInputStream);
         } else if is_noncharacter(code) {
             self.emit_error(ErrorKind::NoncharacterInInputStream);
@@ -550,6 +550,24 @@ where
         }
     }
 
+    fn consume_and_append_to_tag_token_name<F>(&mut self, c: char, f: F)
+    where
+        F: Fn(char) -> bool,
+    {
+        let b = self.buf.clone();
+        let mut buf = b.borrow_mut();
+        let b = self.sub_buf.clone();
+        let mut sub_buf = b.borrow_mut();
+
+        buf.push(c.to_ascii_lowercase());
+        sub_buf.push(c);
+
+        let value = self.input.uncons_while(f);
+
+        buf.push_str(&value.to_ascii_lowercase());
+        sub_buf.push_str(value);
+    }
+
     fn finish_tag_token_name(&mut self) {
         if let Some(
             Token::StartTag {
@@ -601,6 +619,24 @@ where
 
         buf.push(c);
         sub_buf.push(raw_c);
+    }
+
+    fn consume_and_append_to_attribute_token_name<F>(&mut self, c: char, f: F)
+    where
+        F: FnMut(char) -> bool,
+    {
+        let b = self.buf.clone();
+        let mut buf = b.borrow_mut();
+        let b = self.sub_buf.clone();
+        let mut sub_buf = b.borrow_mut();
+
+        buf.push(c.to_ascii_lowercase());
+        sub_buf.push(c);
+
+        let value = self.input.uncons_while(f);
+
+        buf.push_str(&value.to_ascii_lowercase());
+        sub_buf.push_str(value);
     }
 
     fn finish_attribute_token_name(&mut self) {
@@ -672,12 +708,35 @@ where
         }
     }
 
-    #[inline(always)]
-    fn append_to_attribute_token_value_raw(&mut self, raw_c: char) {
+    fn consume_and_append_to_attribute_token_value<F>(&mut self, c: char, f: F)
+    where
+        F: FnMut(char) -> bool,
+    {
+        let b = self.buf.clone();
+        let mut buf = b.borrow_mut();
         let b = self.sub_buf.clone();
         let mut sub_buf = b.borrow_mut();
 
-        sub_buf.push(raw_c);
+        let is_cr = c == '\r';
+
+        if is_cr {
+            buf.push('\n');
+            sub_buf.push(c);
+
+            if self.input.cur() == Some('\n') {
+                self.input.bump();
+
+                sub_buf.push('\n');
+            }
+        } else {
+            buf.push(c);
+            sub_buf.push(c);
+        }
+
+        let value = self.input.uncons_while(f);
+
+        buf.push_str(value);
+        sub_buf.push_str(value);
     }
 
     fn finish_attribute_token_value(&mut self) {
@@ -778,7 +837,10 @@ where
         sub_buf.push(raw_c);
     }
 
-    fn handle_raw_and_append_to_comment_token(&mut self, c: char) {
+    fn consume_and_append_to_comment_token<F>(&mut self, c: char, f: F)
+    where
+        F: Fn(char) -> bool,
+    {
         let b = self.buf.clone();
         let mut buf = b.borrow_mut();
         let b = self.sub_buf.clone();
@@ -799,6 +861,11 @@ where
             buf.push(c);
             sub_buf.push(c);
         }
+
+        let value = self.input.uncons_while(f);
+
+        buf.push_str(value);
+        sub_buf.push_str(value);
     }
 
     fn emit_comment_token(&mut self, raw_end: Option<&str>) {
@@ -1158,7 +1225,7 @@ where
                     // Switch to the before attribute name state.
                     Some(c) if is_spacy(c) => {
                         self.finish_tag_token_name();
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                         self.state = State::BeforeAttributeName;
                     }
                     // U+002F SOLIDUS (/)
@@ -1178,7 +1245,7 @@ where
                     // Append the lowercase version of the current input character (add 0x0020
                     // to the character's code point) to the current tag token's tag name.
                     Some(c) if is_ascii_upper_alpha(c) => {
-                        self.append_to_tag_token_name(c.to_ascii_lowercase(), c);
+                        self.consume_and_append_to_tag_token_name(c, is_ascii_upper_alpha);
                     }
                     // U+0000 NULL
                     // This is an unexpected-null-character parse error. Append a U+FFFD
@@ -1200,7 +1267,17 @@ where
                     // Append the current input character to the current tag token's tag name.
                     Some(c) => {
                         self.validate_input_stream_character(c);
-                        self.append_to_tag_token_name(c, c);
+                        self.consume_and_append_to_tag_token_name(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
+
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control
+                            !is_spacy(c)
+                                && !matches!(c, '/' | '>' | '\x00')
+                                && !is_ascii_upper_alpha(c)
+                        });
                     }
                 }
             }
@@ -1265,7 +1342,7 @@ where
                     // to the before attribute name state. Otherwise, treat it as per the
                     // "anything else" entry below.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
 
                         if self.current_end_tag_token_is_an_appropriate_end_tag_token() {
                             self.finish_tag_token_name();
@@ -1385,7 +1462,7 @@ where
                     // to the before attribute name state. Otherwise, treat it as per the
                     // "anything else" entry below.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
 
                         if self.current_end_tag_token_is_an_appropriate_end_tag_token() {
                             self.finish_tag_token_name();
@@ -1513,7 +1590,7 @@ where
                     // to the before attribute name state. Otherwise, treat it as per the
                     // "anything else" entry below.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
 
                         if self.current_end_tag_token_is_an_appropriate_end_tag_token() {
                             self.finish_tag_token_name();
@@ -1809,7 +1886,7 @@ where
                     // to the before attribute name state. Otherwise, treat it as per the
                     // "anything else" entry below.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
 
                         if self.current_end_tag_token_is_an_appropriate_end_tag_token() {
                             self.finish_tag_token_name();
@@ -2151,7 +2228,7 @@ where
                     // U+0020 SPACE
                     // Ignore the character.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                     }
                     // U+002F SOLIDUS (/)
                     // U+003E GREATER-THAN SIGN (>)
@@ -2200,7 +2277,7 @@ where
                     // Reconsume in the after attribute name state.
                     Some(c) if is_spacy(c) => {
                         self.finish_attribute_token_name();
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                         self.reconsume_in_state(State::AfterAttributeName);
                     }
                     Some('/' | '>') | None => {
@@ -2217,7 +2294,9 @@ where
                     // Append the lowercase version of the current input character (add 0x0020
                     // to the character's code point) to the current attribute's name.
                     Some(c) if is_ascii_upper_alpha(c) => {
-                        self.append_to_attribute_token_name(c.to_ascii_lowercase(), c);
+                        self.consume_and_append_to_attribute_token_name(c, |c| {
+                            is_ascii_upper_alpha(c)
+                        });
                     }
                     // U+0000 NULL
                     // This is an unexpected-null-character parse error. Append a U+FFFD
@@ -2240,8 +2319,17 @@ where
                     // Append the current input character to the current attribute's name.
                     Some(c) => {
                         self.validate_input_stream_character(c);
+                        self.consume_and_append_to_attribute_token_name(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
 
-                        anything_else(self, c);
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control
+                            !is_spacy(c)
+                                && !matches!(c, '/' | '>' | '=' | '\x00' | '"' | '\'' | '<')
+                                && !is_ascii_upper_alpha(c)
+                        });
                     }
                 }
 
@@ -2265,7 +2353,7 @@ where
                     // U+0020 SPACE
                     // Ignore the character.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                     }
                     // U+002F SOLIDUS (/)
                     // Switch to the self-closing start tag state.
@@ -2311,18 +2399,18 @@ where
                     // U+0020 SPACE
                     // Ignore the character.
                     Some(c) if is_spacy(c) => {
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                     }
                     // U+0022 QUOTATION MARK (")
                     // Switch to the attribute value (double-quoted) state.
                     Some(c @ '"') => {
-                        self.append_to_attribute_token_value_raw(c);
+                        self.append_to_attribute_token_value(None, Some(c));
                         self.state = State::AttributeValueDoubleQuoted;
                     }
                     // U+0027 APOSTROPHE (')
                     // Switch to the attribute value (single-quoted) state.
                     Some(c @ '\'') => {
-                        self.append_to_attribute_token_value_raw(c);
+                        self.append_to_attribute_token_value(None, Some(c));
                         self.state = State::AttributeValueSingleQuoted;
                     }
                     // U+003E GREATER-THAN SIGN (>)
@@ -2348,7 +2436,7 @@ where
                     // Switch to the after attribute value (quoted) state.
                     // We set value to support empty attributes (i.e. `attr=""`)
                     Some(c @ '"') => {
-                        self.append_to_attribute_token_value_raw(c);
+                        self.append_to_attribute_token_value(None, Some(c));
                         self.state = State::AfterAttributeValueQuoted;
                     }
                     // U+0026 AMPERSAND (&)
@@ -2377,7 +2465,15 @@ where
                     // Append the current input character to the current attribute's value.
                     Some(c) => {
                         self.validate_input_stream_character(c);
-                        self.append_to_attribute_token_value(Some(c), Some(c));
+                        self.consume_and_append_to_attribute_token_value(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
+
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control, `\r` is in list because of newline normalization
+                            !matches!(c, '"' | '&' | '\x00' | '\r')
+                        });
                     }
                 }
             }
@@ -2389,7 +2485,7 @@ where
                     // Switch to the after attribute value (quoted) state.
                     // We set value to support empty attributes (i.e. `attr=''`)
                     Some(c @ '\'') => {
-                        self.append_to_attribute_token_value_raw(c);
+                        self.append_to_attribute_token_value(None, Some(c));
                         self.state = State::AfterAttributeValueQuoted;
                     }
                     // U+0026 AMPERSAND (&)
@@ -2418,7 +2514,15 @@ where
                     // Append the current input character to the current attribute's value.
                     Some(c) => {
                         self.validate_input_stream_character(c);
-                        self.append_to_attribute_token_value(Some(c), Some(c));
+                        self.consume_and_append_to_attribute_token_value(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
+
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control, `\r` is in list because of newline normalization
+                            !matches!(c, '\'' | '&' | '\x00' | '\r')
+                        });
                     }
                 }
             }
@@ -2437,7 +2541,7 @@ where
                     // Switch to the before attribute name state.
                     Some(c) if is_spacy(c) => {
                         self.finish_attribute_token_value();
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                         self.state = State::BeforeAttributeName;
                     }
                     // U+0026 AMPERSAND (&)
@@ -2487,8 +2591,19 @@ where
                     // Append the current input character to the current attribute's value.
                     Some(c) => {
                         self.validate_input_stream_character(c);
+                        self.consume_and_append_to_attribute_token_value(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
 
-                        anything_else(self, c);
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control, `\r` is in list because of newline normalization
+                            !is_spacy(c)
+                                && !matches!(
+                                    c,
+                                    '&' | '>' | '\x00' | '"' | '\'' | '<' | '=' | '`' | '\r'
+                                )
+                        });
                     }
                 }
             }
@@ -2503,7 +2618,7 @@ where
                     // Switch to the before attribute name state.
                     Some(c) if is_spacy(c) => {
                         self.finish_attribute_token_value();
-                        self.skip_next_lf(c);
+                        self.skip_whitespaces(c);
                         self.state = State::BeforeAttributeName;
                     }
                     // U+002F SOLIDUS (/)
@@ -2607,7 +2722,15 @@ where
                     // Append the current input character to the comment token's data.
                     Some(c) => {
                         self.validate_input_stream_character(c);
-                        self.handle_raw_and_append_to_comment_token(c);
+                        self.consume_and_append_to_comment_token(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
+
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control, `\r` is in list because of newline normalization
+                            !matches!(c, '>' | '\x00' | '\r')
+                        });
                     }
                 }
             }
@@ -2840,7 +2963,15 @@ where
                     // Append the current input character to the comment token's data.
                     Some(c) => {
                         self.validate_input_stream_character(c);
-                        self.handle_raw_and_append_to_comment_token(c);
+                        self.consume_and_append_to_comment_token(c, |c| {
+                            if !is_allowed_character(c) {
+                                return false;
+                            }
+
+                            // List of characters from above to stop consumption and a certain
+                            // branch took control, `\r` is in list because of newline normalization
+                            !matches!(c, '<' | '-' | '\x00' | '\r')
+                        });
                     }
                 }
             }
@@ -4069,7 +4200,6 @@ where
                     // numeric character reference state.
                     Some(c @ '#') => {
                         self.temporary_buffer.push(c);
-
                         self.state = State::NumericCharacterReference;
                     }
                     // Anything else
@@ -4540,7 +4670,7 @@ where
     }
 
     #[inline(always)]
-    fn skip_next_lf(&mut self, c: char) {
+    fn skip_whitespaces(&mut self, c: char) {
         if c == '\r' && self.input.cur() == Some('\n') {
             self.input.bump();
         }
@@ -4641,4 +4771,20 @@ fn is_ascii_lower_alpha(c: char) -> bool {
 #[inline(always)]
 fn is_ascii_alpha(c: char) -> bool {
     is_ascii_upper_alpha(c) || is_ascii_lower_alpha(c)
+}
+
+#[inline(always)]
+fn is_allowed_control_character(c: u32) -> bool {
+    c != 0x00 && is_control(c)
+}
+
+#[inline(always)]
+fn is_allowed_character(c: char) -> bool {
+    let c = c as u32;
+
+    if is_surrogate(c) || is_allowed_control_character(c) || is_noncharacter(c) {
+        return false;
+    }
+
+    return true;
 }