--- a/crates/swc_html_parser/tests/recovery/doctype/newline/dom.rust-debug
+++ b/crates/swc_html_parser/tests/recovery/doctype/newline/dom.rust-debug
@@ -1,0 +1,6 @@
+| <!DOCTYPE html "-//W3C//DTD HTML 4.01
+
+Transitional//EN" "">
+| <html>
+|   <head>
+|   <body>