--- a/crates/swc_html_parser/tests/recovery/doctype/newline/input.html
+++ b/crates/swc_html_parser/tests/recovery/doctype/newline/input.html
@@ -1,0 +1,3 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01
+
+Transitional//EN">