--- a/crates/swc_html_parser/tests/recovery/doctype/newline/output.json
+++ b/crates/swc_html_parser/tests/recovery/doctype/newline/output.json
@@ -1,0 +1,66 @@
+{
+  "type": "Document",
+  "span": {
+    "start": 1,
+    "end": 65,
+    "ctxt": 0
+  },
+  "mode": "no-quirks",
+  "children": [
+    {
+      "type": "DocumentType",
+      "span": {
+        "start": 1,
+        "end": 65,
+        "ctxt": 0
+      },
+      "name": "html",
+      "publicId": "-//W3C//DTD HTML 4.01\n\nTransitional//EN",
+      "systemId": null,
+      "raw": "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.01\n\nTransitional//EN\">"
+    },
+    {
+      "type": "Element",
+      "span": {
+        "start": 0,
+        "end": 0,
+        "ctxt": 0
+      },
+      "tagName": "html",
+      "namespace": "http://www.w3.org/1999/xhtml",
+      "attributes": [],
+      "children": [
+        {
+          "type": "Element",
+          "span": {
+            "start": 0,
+            "end": 0,
+            "ctxt": 0
+          },
+          "tagName": "head",
+          "namespace": "http://www.w3.org/1999/xhtml",
+          "attributes": [],
+          "children": [],
+          "content": null,
+          "isSelfClosing": false
+        },
+        {
+          "type": "Element",
+          "span": {
+            "start": 0,
+            "end": 0,
+            "ctxt": 0
+          },
+          "tagName": "body",
+          "namespace": "http://www.w3.org/1999/xhtml",
+          "attributes": [],
+          "children": [],
+          "content": null,
+          "isSelfClosing": false
+        }
+      ],
+      "content": null,
+      "isSelfClosing": false
+    }
+  ]
+}