--- a/crates/swc_html_parser/tests/recovery/doctype/newline/output.stderr
+++ b/crates/swc_html_parser/tests/recovery/doctype/newline/output.stderr
@@ -1,0 +1,7 @@
+
+  x Non conforming doctype
+   ,-[$DIR/tests/recovery/doctype/newline/input.html:1:1]
+ 1 | ,-> <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01
+ 2 | |   
+ 3 | `-> Transitional//EN">
+   `----