--- a/crates/swc_html_parser/tests/recovery/doctype/newline/span.rust-debug
+++ b/crates/swc_html_parser/tests/recovery/doctype/newline/span.rust-debug
@@ -1,0 +1,25 @@
+
+  x Document
+   ,-[$DIR/tests/recovery/doctype/newline/input.html:1:1]
+ 1 | ,-> <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01
+ 2 | |   
+ 3 | `-> Transitional//EN">
+   `----
+
+  x Child
+   ,-[$DIR/tests/recovery/doctype/newline/input.html:1:1]
+ 1 | ,-> <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01
+ 2 | |   
+ 3 | `-> Transitional//EN">
+   `----
+
+  x DocumentType
+   ,-[$DIR/tests/recovery/doctype/newline/input.html:1:1]
+ 1 | ,-> <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01
+ 2 | |   
+ 3 | `-> Transitional//EN">
+   `----
+
+  x Child
+
+  x Element