<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

- Also small bug fix for `swc_html_codegen`

perf:
```
html/parser_document/css_2021_spec                                                                            
                        time:   [23.669 ms 23.763 ms 23.871 ms]
                        change: [-4.2093% -3.4579% -2.7503%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

Benchmarking html/parser_document/github_com_17_05_2022: Collecting 100 samples in estimated 5.1213 s (600 iterati                                                                                                                  html/parser_document/github_com_17_05_2022                        
                        time:   [8.2812 ms 8.3389 ms 8.4115 ms]
                        change: [-11.642% -10.727% -9.6697%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe

Benchmarking html/parser_document/stackoverflow_com_17_05_2022: Collecting 100 samples in estimated 5.1562 s (700                                                                                                                   html/parser_document/stackoverflow_com_17_05_2022                        
                        time:   [7.2896 ms 7.3277 ms 7.3745 ms]
                        change: [-10.511% -9.7884% -9.0822%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe

```

**BREAKING CHANGE:**

No

**Related issue (if exists):**

No